Check of correct sensor ranges measured
Calls the total stream function via Bluetooth during a given period of time. The values obtained from the sensors are compared with the expected value ranges and a warning is given when only one of the them is out of range. This is very helpful in order to check data rate of the Bluetooth communication, corrupted values and firmware and hardware issues.